+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.2.5 (Upcoming Release)
 
 - Add new tests for CPM
+- Feed an initial feasible solution (with North-West Corner Method) to a Transportation Problem. That makes it 10x faster
 
 
 ### 0.2.4