+++ CHANGELOG.md
@@ -1,9 +1,12 @@
-### 0.2.5 (Upcoming Release)
+### 0.2.6 (Upcoming Release)
+
+
+### 0.2.5 
 
 - Add new tests for CPM
 - Feed an initial feasible solution (with North-West Corner Method) to a Transportation Problem. That makes it 10x faster
 - Implement least cost method for constructing a feasible initial solution to the transportation problems.
-
+- Update documentation
 
 ### 0.2.4